+++ api/v1/credentials/read.php
@@ -37,8 +37,8 @@ if ($sql && mysqli_num_rows($sql) > 0) {
 
     $row = array();
     while ($row = mysqli_fetch_assoc($sql)) {
-        $row['credential_username'] = apiDecryptCredentialEntry($row['credential_username'], $api_key_decrypt_hash, $api_key_decrypt_password);
-        $row['credential_password'] = apiDecryptCredentialEntry($row['credential_password'], $api_key_decrypt_hash, $api_key_decrypt_password);
+        $row['credential_username'] = apiDecryptCredentialEntry($row['credential_username'], $api_key_row ?? $api_key_decrypt_hash, $api_key_decrypt_password);
+        $row['credential_password'] = apiDecryptCredentialEntry($row['credential_password'], $api_key_row ?? $api_key_decrypt_hash, $api_key_decrypt_password);
         $return_arr['data'][] = $row;
     }